More cleanup of deprecated systems
Missed a few references and pieces of code, and also:

(1) We still allowed ReplaceGrammar._toTapeNames to hold a list of strings, but this hasn't actually been allowed for a long time.  Changing this to `ReplaceGrammar._toTapeName: string` which simplifies some of the logic that uses it.

(2) `JoinRule` was a bit of an awkward name; it referenced a `rule` (new-style implicit-tape rule) vs. `replace` (old-style explicit-tape rule) distinction that, with the elimination of JoinReplace, now no longer exists.  It's been renamed `ReplaceBlock` because that's what it corresponds to in the syntax.

(3) Unrelatedly, found an bug in the simplification of MatchFrom where, when constructing a cascade of them, we only simplified the outer one.  Fixed this.