Filter errors at compile time rather than runtime
This commit addresses a major issue that I thought we had addressed a long time ago... and hadn't.

Background: FilterExprs arose years ago with the question "If someone says `equals gloss: foo` and there's no 'gloss' tape, what do we do?  Do we add 'gloss:foo' to every result?  Or is that the empty set?" We decided it was and error and should result in the empty set, and achieved this with a special expression FilterExpr that enforced this.

But really, this isn't the appropriate place to handle this kind of error; this is a *programmer* error and needs to be handled at compile time so the programmer can see the error and fix it.  Also, we now have the principle of "Try not to return the empty set because that's impossible for the programmer to debug".

We actually removed FilterGrammars/Exprs from equals/starts/ends/contains some time ago in favor of Joins, for just this reason.. but looking through everything, we never actually wrote any code to catch the tape error :(  That's a big problem; filtering on a tape that doesn't exist is a very common error, and with starts/ends it can cause exponential generation.

This commit fixes this so that the missing tape error is caught and flagged, and the result is just the unfiltered grammar rather than the empty set.

I also implemented this for "test:" blocks.  (There's no exponential crisis here, but it was still less than ideal.)  Previously, tests were implemented as Filters, so if you asked for a test with "gloss" for a grammar that had no "gloss" in the first place, the test just failed.  This was okay but not great, because it's a different kind of error than just a test failing.  (And this would show up *passing* in a "testnot:" block, which also isn't great.)  Now it works more like equals/starts/etc.: the error is flagged to the programmer at compile time, and the erroneous tests don't execute at all.

This was the last remaining use of FilterGrammar/Expr outside of our own internal testing, and most of those tests were recapitulations of existing tests using Join, so I eliminated FilterGrammar/Expr.

There's now a new FilterGrammar in its place, but it's doing something entirely different.  It's just there to get the aforementioned equals/starts/etc. error messages working.  It exists briefly in between creating Grammars and the point at which it can flag the tape-mismatch error, then if nothing goes wrong it's replaced with a JoinGrammar.