Refactored the Replace tests (testGrammarReplace) as per commit 87740a1; testGenerate enhancements.
Extended the restriction argument to Interpreter.generate() to accept a list of
StringDicts or a StringDict. The list is interpreted as an alternation of
sequences, with each sequence represented by a StringDict.

Added the new restriction argument to testGenerate. This allows the tests in
testGrammarReplace that previously used testParseMultiple to use testGenerate
(via testGrammar) instead.

Added test timing within testGenerate and testParseMultiple (using timeIt). The
timings reporting is triggered by VERBOSE_TEST_L2 being set to true. A shortDesc
argument has also been added to testGenerate to facilitate test identification.

Added a withVocab convenience function in testGrammarUtil because I was defining
it in many test suite files. This convenience function perhaps belongs in
grammars.ts, but I put it in testGrammarUtil.ts for now.

Some tests in testGrammarReplace took a long time to run; timeIt revealed which
ones. It happens with tests involving longer matches without the replacement
being restricted to the beginning of the string. It appears that there just are
that many states that legitimately have to be checked. Tests 2a-b exhibit the
problems but run in under 2 seconds each, so I left them enabled. Tests 6a-c
require considerably longer to run (4-20s), so I set them to skip generation by
default.