Fixing critical Cursor and Correspond issues
(1) I had always talked about Cursors as if, upon delta'ing out on their associated tape, they ceased to exist.  But this isn't actually what happens in the code!  Instead, they actually stick around until the very end, they're only actually destroyed when all that's left is outputs and everything collapses into one OutputExpr during simplify().

This is an efficiency problem because they continue to make delta/deriv queries on tapes that we know are finished, and delta/deriv are expensive.

Now there's a "finished" boolean in Cursors.  They still stay in place (they're holding the outputs) but they no longer make queries, they only send the .forward() call to their children.

Eventually finished Cursors should probably be their own kind of thing, but it's not a big deal at this moment.

(2a) Fixing this revealed a deeper bug.  I had been wondering why all of our tests went green -- especially the outstanding "something goes wrong with replace when you query the tapes in an unexpected order" bug.

That had pretty obviously been an issue with Corresponds not yet working in a "priority-independent" fashion.  But I hadn't touched that, why were they suddenly working?  It turns out that Cursor sending those otherwise-useless delta queries was making the CorrespondExprs work.

Background: CorrespondExprs are kind of strange compared to other exprs, the basic idea is that they forbid themselves to delta-out/simpify unless a particular condition is met (that we've emitted as many characters on the output tape as the input).  Until then, they stick around in the tree.  In normal circumstances (i.e. when we query their tapes in the expected order, fromTape before toTape), this condition is eventually met.  But when we query toTape before fromTape, this changes the way we compare lengths (we basically get an off-by-one error), the condition is never met, it doesn't simplify, and we throw our oops-didn't-simplify exception.  (That off-by-one error was very hard to address because adjusting for the off-by-one depends on information CorrespondExpr ideally shouldn't have.)

Why did this suddenly "fix" itself?  It turns out by sending a SECOND delta query to toTape, the condition can be met and the CorrespondExpr simplifies.  Ordinarily, of course, there will never be a second delta, but because of the bugged CursorExpr we were sending deltas on finished tapes all the time.

Obviously this is not an appropriate solution.

We could actually leave CorrespondExpr bugged like this, because we don't ever use the bugged configuration, but the brittleness of CorrespondExpr has been a thorn in our side for a while and so this is more motivation to fix it.

(2b) The new CorrespondExpr doesn't count characters anymore.  Instead, it has a different way of operating.  Basically, only a delta on the fromTape can successfully get rid of a CorrespondExpr.  It DOESN'T permit delta-ing out on the toTape at all, the only way to delta out on the toTape is if the fromTape already delta'd out (resulting in there not being a CorrespondExpr anymore to prevent toTape delta-ing out).

Unusually compared to most exprs, it doesn't simplify itself down to epsilon when its child is epsilon.  That's not a way to get rid of it, as before only a successful deltaing-out on fromTape gets rid of it.  (That will, of course, happen soon if the child is epsilon.  But it doesn't happen right away.  Simplifying too early leads to some incorrect results.)

(2c) We were never directly testing CorrespondExpr, because there was no corresponding CorrespondGrammar.  But we'll eventually need one anyway, so I made it, and then wrote some tests.