Cleanup and more tests for new Correspond
(1) CorrespondExpr simplification was still happening in constructCorrespond rather than `.simplify()`, now it's where it belongs.

(2) Some more small tests.  In particular I wanted to make sure that we weren't introducing an inefficency (albeit a minor 1-cycle one) when fromTape and toTape were the *same* length.

When we query the tapes in the expected order, there's no inefficency.  When we query the tapes in the backwards order, there is a 1-cycle delay, but not an unexpected one.