Simplifying sampling and testing for it
We used to have an issue when sampling from a grammar, that upon finding one output we would yield it and stop, even if it was a prefix of another output.  (E.g., if `{t1:hell}` was a valid output, we would never generate `{t1:hello}`, because to get there we'd get to the former first and that would be the one we sample.)

To get around this, instead of yielding `{t1:hell}` we'd put this in a buffer and give the generator a few more rounds to let it potentially find additional outputs, then yield randomly from that buffer.

I just noticed this problem actually doesn't occur anymore, and hasn't from some time.  The reason is because we don't immediately yield `{t1:hell}` anyway, it goes on the stack and we yield it once we find it on the stack.  Since randomization causes the additions to the stack to be pushed in random order, this might or might not be the state we visit next.  We might visit the state that will become `{t1:hello}` next instead.

So anyway, the whole buffering solution had become unnecessary, and I've removed it, which simplifies the generation loop further.

I also added a few sampling tests, to confirm that this is indeed what is happening.  (We used to have sampling tests but that was a long time ago and they've been lost.)  These are in their own subfolder because we don't want to run them 100% of the time, just occasionally.