Re-fix for priority
The fix I made for Cursor priority last week fixed things in the wrong place.  The priority calculation wasn't backwards, I was nesting Cursors in the wrong order in the convenience function `Cursor(tapes, child)`.  This made the use of that function counterintuitive, since the priority happened in the opposite direction than what you'd expect.  This commit makes it work the way you'd expect.