Investigating the sampling slowdown
I've basically figured out what's slowing down sampling in this branch.  It's way simpler than I anticipated.

In both the `main` and `derivUnify` branches, the sampler is basically doing its work after 1 "tick" of the generator loop: it takes all the states it finds after one tick, shuffles them, then puts them on the stack.

But waaaaay more is happening every tick in `derivUnify`; we're querying on EVERY tape rather than just ONE tape.  When a grammar has a lot of possible forms and a lot of possible tapes, and not a lot of rules, almost everything is happening in that very first tick!  (E.g., our stripped down copy of the Oneida grammar has about 2500 forms, and by the very first tick we've got 2500 states... and then we're shuffling them, which is expensive.)

So sampling was taking almost as long (or longer) as generation.  (One of the reasons I didn't notice this is I hadn't noticed that generation has become quite fast.)

Luckily, `.forward()` is a generator -- we can ask for N things off of it, then resume later.  I've started to implement this, but I haven't quite gotten it right (appear to be losing some forms).  But I wanted to get this committed while I had a partial solution.