Fixing missing outputs in last commit
The problem of missing outputs in the last commit ended up being easy.  When refactoring `iterTake()` to also return the unfinished generator (for cases where we need to resume), I was accidentally taking one more yield and not adding it to the results.  That's fixed.

So both generation and sampling is fast now.

However, sampling still isn't completely fixed -- it's wildly un-uniform and biased to sample from the first few yielded forms.  That's because we're taking the first N states, then a "remainder" `GenExpr` that represents all the remaining ones, and shuffling those.  That means it's an N/(N+1) chance that our sample comes from those N, and 1(N+1) chance that it comes from the remainder.

That bias towards early forms is fine (I mean, we need that, if it didn't have that bias it'd be slow again!), the bigger issue is that our ordering is deterministic.  There's nothing built into the `forward/delta/deriv` system that's non-deterministic, all of the non-determinism is `generate` loop itself.  So some aspect of the generation algorithm is going to have to become sensitive to `env.random`.