Moving randomization from generate() to .forward()
In the previous commit, I mentioned a dilemma.

    (a) If we wait until we have all the possible next-states resulting from `.forward()` in order to randomize them, we'll be waiting a long time, and it also takes a long time to randomize.

    (b) But if we grab some number N from `.forward()` and randomize those, we end up with a very non-uniform sampling strategy.

This dilemma only arises if `.forward()` is determinisic, which it was.

So this commit makes `.forward()` check `env.opt.random` and, if it's true, shuffle the order of its own-tape results.  (That is, if the Cursor is relative to `t1`, shuffling the results of d_t1 and D_t1.)  It's not shuffling all of the results including the child's forwards, that would just put us back in the position of shuffling all the results.

This is basically the same shuffling that `generate()` used to do, back when `generate()` was the one calling deriv and delta directly (or, later, through PriorityExpr).  Either way, it only called those w.r.t. one tape at a time, so when it shuffled it was only shuffling on one tape.

So you can see this as "We moved a lot of responsibilities from `generate()` into `CursorExpr.forward()`, but left randomization behind, and thus changed its scope: from randomizing one-tape results to randomizing all-tape results!  Now it's restored to its proper logical position.

This also simplifies the generation loop once again.  We used to have a variable `nexts`.  Rather than push results directly onto the search stack, we first pushed them to a `nexts` buffer, then (if random) shuffled them before pushing them to the stack.  (This was in order to avoid shuffling the whole stack every time!)  Now the results already arrive in a shuffled order, so there's no need for this intermediate buffer.

(Nor is there a need for the trick introduced last time, to shuffle the `remainder: GenExpr` into `nexts`.  That was just a stopgap solution.  It's fine now for `remainder` to sit under the rest of the results on the stack.  We really only had to shuffle them to make it theoretically *possible* to sample from later results; we have a better solution now.)