Randomizing multi-output results
The recent changes lost us the ability to sample properly when one OutputExpr corresponds to multiple results.  That's not the usual case, but it's possible (e.g. when we have an alternation of Cursors).

Luckily this is simple to restore, by shuffling the results of `prev.getOutputs()` before yielding them.

(This is a little inefficient, since we're shuffling an array and then in practice only grabbing the first thing... but this isn't a usual occurence anyway.)