Experiments in quick sampling
These were a bunch of small experiments in more efficient sampling, suggesting we can get a very large improvement in sampling speed by taking advantage of our generators.

(Background: We always wrote derivs as generators with this purpose in mind, but over time a lot of them didn't get used as generators anymore: bit by bit they started to get used as arrays.  In `disjoin()`, in `Cursor.forward()`, in `UnionExpr.deriv()`, etc., we were effectively going `[...gen]` and thus losing the ability to get the first item without going throug the rest.)

(1a) `disjoin()` sometimes really does need to go through the whole list, like when we're disjoining for the purpose of negation, but a lot of times it's just to avoid going through duplicates.  So I made it sensitive to an `.obligatory` param and also to whether the sampling is random.

(1b) `UnionExpr` now only does its collect-and-make-a-new-union when it's not random, otherwise it goes back to its old behavior.  (This may cause some issues but in the short term this works as an illustration.)

(1c) `Cursor.forward()` needed to collect all the results of `deriv` on its relevant tape to shuffle them for randomization.  This is a conflict: when randomization is on, we don't want to collect all the results because it's wasted effort, but also we need to collect them all for randomization :sob:.

Anyway, experimentally turning this off gets us a big boost that we can't ignore, suggestion we need a better means of randomization than this shuffle.

(2) You might think as a first attempt, "This is a matter of shuffling all the UnionExprs"  or "This is a matter of yielding randomly from UnionExprs" but neither of these is sufficient, not with our current codebase.  There are other places where multiple alternatives are considered where a UnionExpr is nary to be seen, like in `ConcatExpr.deriv`.

The fundamental problem in `ConcatExpr.deriv` and generators like it is that they yield their results in a particular order: first they yield the ones that result from derivving the first child, and then they yield ones that result from delta'ing the first child and derivving the second.  The former will always be generated before the latter, and will be sampled first.

So this also contains an initial experiment in keeping both sides to be generators all the way through, which gives us the possibility of choosing from those generators at random when it's time to yield from one.  (I didn't do this yet, but it's now a possibility.)

(3) Oooops, that whole thing about sending too many deltas on finished tapes, that I fixed a few commits ago?  I obliterated it later when I fixed something else.  This restores it.  Note to self that even if I abandon this branch, I should port this fix back to main.