Shuffling generators instead of results
The previous commit showed that it was possible for sampling to be fast... but it wasn't actually random!  We had gotten rid of shuffling results, so we were just sampling the same output every time.

There was an easy solution to this:

(1) There were multiple places, mentioned in the last commit, where we were explictly or implicity yield*ing from a sequence of generators (such as the deriv generators of UnionExpr's children, the deriv generators corresponding to the two types of ConcatExpr derivs, the delta and deriv generators of `CursorExpr.forward`, etc.).  What we need is just a generator `iterRandom` that takes that sequence of generators, shuffles the generator sequence without executing them, and then yields from the shuffled sequence.

Of course, if any of those component generators isn't random, we'll fail to sample things, so the key is to make sure that the component generators have this property too.  If they can return multiple results, they should be using `iterRandom` when `env.opt.random` is true.  (That's why we had to refactor `ConcatExpr.deriv` last time, so that it could use `iterRandom` this time.)

(2) I think the only thing remaining that has this property that doesn't use `iterRandom` is NegationExpr, so that'll still be sampled in a fixed order and thus some results will be shadowed.  It shouldn't actually come up when sampling from actual grammars but still

    TODO fix sampling in negation.

(3) Other important note, obviously the distribution of results from shuffling the generators is not uniform.  If you do this to [generatorOf1Thing, generatorOf1000Things], you'll generate that 1 thing half the time!  But really, we've never had uniform sampling anyway.  At some point we have to choose whether to look down the derivational path starting `t1:a` vs. `t1:b` without knowing in advance how many outputs are down the two paths.  `t1:a` might have 1 output, `t1:b` might have a thousand, but we have to make that decision now.