Fixing sampling for Dot, DotStar, & Not
In the last commit I hypothesized that Negation was the only remaining deriv that yielded multiple times in a fixed order.  Actually, Dot and DotStar did, too: they were ordered by the order of the vocabulary.

This commit makes all three of these yield randomly when env.opt.random is true.

(`NegationExpr.deriv` could have been difficult, but because Negation always disjoins its results anyway, we might as well collect and shuffle them: easy.)

I found one more that does this, PreTapeExpr, but that's a bit of a can of worms, I'd rather wait until we can replace those with Cursors.