Fixed blown stack with nested GenExpr
When sampling, we want to be able to explore one yield from the grammar, without throwing away the remainder in case that one yield doesn't pan out.

So a remainder, itself, is something that we have to be able to keep on the stack for the next time.  This is a big annoying, type-wise, because the stack is otherwise `Expr[]` and remainders are `Gen<[boolean,Expr]>`.  `(Expr|Gen<[boolean,Expr]>)[]` is a nasty type, those are two very different kinds of things.

So my quick solution was to wrap the remainder in its own Expr, `GenExpr`, which did nothing but yield from the remainder.  That made them all Exprs and so the stack could remain `Expr[]`.

Turns out this was a mistake, it was actually possible for big grammars to end up with recursively nested GenExprs and then blow the stack with a call to forward.  (Unfortunately, I couldn't construct a test grammar that replicates this, it's something that only seems to happen in big grammars... exactly WHAT kinds of bigness are necessary to replicate, I didn't manage to figure out.)

But in any case, it's clear that just wrapping remainders in GenExpr has got to go, so I bit the bullet and accepted that the stack has to be that complex type.  This addresses the issue.

Side change: While looking through sampling logs, it became clear that we were wasting a bunch of time yielding nulls from deltas, so now `forward()` doesn't yield nulls.  (Or rather, this probably took a negligible amount of runtime, but visually it makes for a lot of clutter in the verbose-debug logs.)