Fixing random result order in generation
Oops, when making results non-deterministic for sampling, I over did it.  In a few places I used `iterRandom()` even when `env.opt.random` was false.

The resulting non-deterministic order doesn't lead to wrong answers, but it messed with a few tests whose descriptions involved stringification of results.  (Tests effectively run twice, once to collect the tests and the second to check the results.  If the description changes in between these two runs, the test results can't be associated with the original test.)