More efficient sampling
When sampling, we were doing a lot of shuffleArray, but it's not like we need the entire array shuffled when we're only yielding (in the end) one thing from it.

If you're only drawing one card, you don't need to shuffle the deck, you just need to "cut" it.  So that's what we do now, we just choose a random offset into the array and iterate through the array modulo its length.