Consistent internal structure/tapes for Replace
I noticed we actually had three different internal structures for replacements.

    (1) In testGrammarReplace, the input grammars were inherently on t1 (e.g. they're `Lit("t1", "e")` and outputs on t2.

    (2) In testGrammarReplaceBlock, the input grammars were inherently on tape $i and outputs on $o.

    (3) In actual source-built grammars, the input and output grammars were on $t by default (except in some special circumstances), and wrapped in SingleTapeGrammars to adapt them to $i or $o as appropriate.

This made me a little uneasy -- in practice these come out as pretty much the same thing, but they're subtly different and could drift apart semantically.  Also, supporting this variety requires a more complex ReplaceGrammar (e.g. the long-running complication about inferring which tapes need to be negated).

This commit makes the `Replace()` convenience function construct the same structure as in source-built grammars; now everything works like (3).  Meanwhile it removes the complexity/flexibility of ReplaceGrammar to handle arbitrary tapes behind the scenes.  Now the input tape is REPLACE_INPUT_TAPE and the output REPLACE_OUTPUT_TAPE and that's that.

(Actually, much of that flexibility still remains in practice, because SingleTapeGrammar itself is flexible about it's children's grammars.  One of the reasons it was complex to reason about ReplaceBlock structures was that all three of ReplaceBlockGrammar, ReplaceGrammar, and SingleTapeGrammar were flexible about child tapes in different ways.  When SingleTapeGrammar entered the pictures, the flexibility of ReplaceGrammar became redundant.)

I committed this to a separate branch because it changes almost every line of testGrammarReplace (every t1 is now $i, etc.), and didn't want to get in the way of other refactorings Darlene might be doing.