New docusaurus v2 site
We had been using docusaurus v1, but that's no longer maintained and adds a lot of vulnerabilities to our dependency analysis.

Since we didn't have all that much content in it, I decided not to try to port the old site.  (It can't 100% be automated, there's extra fiddly stuff to do.)  So I started a new docusaurus v2 site and just manually moved some of the old files in.