Moved docs out of packages
The documentation is a package in itself (i.e. you can build it), but unlike the other packages it's not something that anyone else will ever need to build.  When you build gramble, check its dependencies and vulnerabilities, etc., includng the docs package is wasted time and a ton of warnings (because the docusaurus team isn't very concerned about vulnerabilities).

So this moves it out of packages into its own directory, where we can build it when necessary but isn't part of the general bootstrap.