CLI install fixes for Windows and QoL improvements
(1) Introduced command-line syntax in the README.

(2) Figured out that we CAN install the symlink on Windows, the user just needs elevated/admin privileges to do so.  Mentioned this in the README too.

(3) In `run-locally`, we were executing directly from `commandLine.ts`, which wasn't working on my legacy Windows machine for some reason.  But it does work when we execute the transpiled `bin/index.js` so now that's what we do.

(4) A few months ago we got rid of default symbols within the language itself, because they open up a semantic can of worms.  But this meant that you MUST use the `-s` or `--symbol` argument on the command line, and to do that you have to understand about fully qualified names.  (Even I make mistakes, I try to refer to them by local names.)

Now, if you don't specify the symbol name, it defaults to `<source>.all`.  But note that this isn't a convention of the Gramble language itself, it's not like before where the symbol "" was actually resolved to the relevant grammar.  This is just a command-line convention and only exists there.

(5) Made an `examples` folder and put in a hello-world example, so that people can verify that their install completed without having to write their own gramble program.