Updated repository references and bumped version number to 1.0.0 in JSON files.
The repo references now refer to the NRC GitHub repository, not Pat's personal
one.

Also added various optional OS-specific @nx/nx-* packages in the
package-lock.json file, which are needed for lerna to install on these OSes
(only @nx/nx-win32-x64-msvc was listed previously).

Also updated @types/chai.

Still to be done: Rationalize the dependency requirements between gramble-cli
and parser packages to a common set for common packages, such as typescript,
chai, mocha, node, etc.

A second, as yet unsolved, problem is that "npm install" must be run TWICE
before "npm run build" to install the dependencies, but for some reason it
tries to build (i.e. run tsc) on gramble-cli before the parser package, but
gramble-cli is dependent on the parser, so it fails the first time, then
wipes clean all the installed node_modules. Running a second time succeeds
because the necessary file dependency is satisfied.