Rationalized the dependencies between gramble-cli, parser, overall packages.
Also updated some dependencies to the latest versions.

To the extent possible, we have rationalized the dependency requirements between
packages/gramble-cli, packages/parser, and packages to a common set for common
packages, such as typescript, chai, mocha, node, etc.

Unfortunately, gramble-cli is still dependent on an old version of typescript
(3.5.3) and ts-node (8.10.2) because compilation fails with more recent versions.
We should update the gramble-cli source to build with typescript 5.2.2.