Fixed the 'npm install' process to build the packages in the correct order.
Added --build option to tsc command to make it build its dependencies (i.e. to
build the referenced projects first). Now 'npm install' builds the parser before
gramble-cli.

Making this change requires the two packages to be built with the same typescript
version. Thus we also fixed commandLine.ts to build with Typescript 5.2.2. All
this required was correcting the type for the options property of the Command
iterface.

We also changed the compiler options in tsconfig.json for gramble-cli to be pretty
much the same as for parser.

Finally, we updated the versions specified for typescript, @types/node, and ts-node
in the gramble-cli package.json to be current (i.e. the same as parser and/or the
overall package.json), with the corresponding changes applied to package-lock.json.
Now, when we build, no node-modules folders (only dist) are created under the
parser and gramble-cli packages.