Refactoring atomicity calculation
The atomicity optimization treats every possible string on a tape as an atomic unit (e.g. so that "1SG.SUBJ" is a single token in the vocab, rather than as ["1", "S", "G", ".", "S", etc.].)  Determining whether we can do that safely requires examining the grammar structure to see whether on material on that tape is BOTH concatenated and joined at some point, or whether there are operators like Dot or Replace that assume character-level processing.

This calculation was threaded through the Grammar hierarchy as a method, but the calculation in question wasn't well named ("getTapeClass") and in any case it was a good candidate to factor it out into its own function, because we only call it once and it doesn't really interact in a complex way with other functions.

So now it's a pattern-matching function of its own, named `determineAtomicity()` in the "passes" directory.  (Yes, it's not a pass exactly, it doesn't return a new tree, but it's pass-y enough for now.  At some point it'd be good to make it into a pass that returns the optimized tree, but that won't work at the moment -- we don't currently do Passes to make changes w.r.t. a particular query.)