Reorganizing opts & threading through passes
We weren't threading atomicity through passes -- we weren't threading ANY options through passes, actually, except verbosity.  This does that, and reorganizes what options you can set where and how they're kept.

(1) GenOptions was originally for options relevant to one pass of generation.  But we don't really need the ability to reset (say) maxRecursion or verbose between different runs of the generator.  (Some time ago, we decided verbosity was a property of the Interpreter anyway.)  Really the only thing that actually needs to vary between different runs of the generator was randomness... and that's not really an "option", it's not a switch the client should have access to.

(1a) `GenOptions` => `Options`.  While the generator needs this information, they're not specific to generation anymore, and other parts of the system need them too.

(1b) Randomness isn't a member of options anymore.  It's a parameter to `generate()` again, and a direct member of `DerivEnv`, not a member of `DerivEnv.opt`.

(2) When constructing an Interpreter (directly or via static functions like fromSheet), you can pass in a `Partial<Options>`.  Any unspecified options will be set to defaults, and the result is stored in the interpreter object.

(2a) `verbose` is part of Options, so passing in the partial replaces the ability to directly pass in `verbose`, e.g.:

    FORMERLY: Interpreter.fromSheet(devEnv, sheetname, VERBOSE_TIME)
    NOW:      Interpreter.fromSheet(devEnv, sheetname, {verbose: VERBOSE_TIME})

(3) The same as DerivEnv has a member `.opt: Options`, so does PassEnv now.