Small env/opt refactoring/fixes
(1) constructExpr wasn't actually using its tapeNS parameter for anything.  (RenameGrammar.constructExpr was modifying it... but then nothing else used that value.)

(2) DerivEnv and PassEnv both inherit from Env now.  There's nothing much in Env right now, just the `opt` member, but it'd be good to move some of the logging functionality there.

(3) infinityProtection was taking a now-extraneous maxChars parameter, but now that PassEnv has an `opt` member, it already has access to that value through the `env` parameter.

(4) Because PassEnv didn't have `opt`, and executeTests is a Pass, unit test execution wasn't using the client-supplied options -- it was making a fresh Options object and using that.  That was always wrong, but with our new understanding that things like maxChars and maxRecursion are properties of the *interpreter* rather than then execution run, it's even more obviously wrong.  So this commit fixes that.