Threading env where it's needed for LTR/RTL
To move the LTR/RTL option into Options, we need to thread an Env anywhere the global DIRECTION_RTL is used, and anywhere that uses this.  This includes:

    * Expr.simplify()
    * Expr.getOutputs()
    * constructX for almost all X

(There are a few constructX where this wasn't needed.  I could have put env there for consistency, but constructX is a pretty heterogeneous collection of functions anyway.)