Making not-contains test work RTL
The optimal expression of not-contains differs between LTR and RTL, so the test for it in testGrammarNegation fails when directionLTR==false.

Instead of just turning it off, I made a parallel test for the RTL case and switch them depending on `opt.directionLTR`.

At some point we'll want to express both, but LTR isn't yet a param to testGrammar.