Test functions allow directionLTR & other Options
To test LTR vs RTL, we used to have to go change the global const DIRECTION_LTR.  In the last commit, this changed to switching based on the value `opt.directionLTR`.

...But since that `opt` was just a new instance of the default options anyway, in practice we were doing the same thing, we still had to go in and change the global default to test both.

This commit lets testGrammar() take any of the Options params -- like where we used to only be able to specify `verbose` and `maxRecursion`, you can now use any of the Options.  The public interface of testGrammar hasn't changed, though, because the Options are added by type intersection.  (That is, you don't have to change anything from `verbose: VERBOSE_DEBUG` to like `opt: {verbose: VERBOSE_DEBUG}`, it's still just `verbose: VERBOSE_DEBUG`.  It's just you can now do that with any Option.)