Small fixes/refactoring of test options
(1) Somehow we ended up with two default max recursions (one defined as DEFAULT_MAX_RECURSION in testUtils, the other as a member of DEFAULT_OPTIONS, formerly GenOptions)... and at some point they drifted to have different values, oops.  I moved DEFAULT_MAX_RECURSION to utils (so that it's alongside DEFAULT_MAX_CHARS) and used that to set the value in DEFAULT_OPTIONS.

(2) Rather than define GrammarTestAux as an intersection of `{stuff} & Options`, made it extend Options on the example of GrammarTest.  I think it ends up the same underneath but extension syntax is more readable.

(3) Oops, we had maxRecursion in the GrammarTestAux twice (once on its own, once inherited from Options).