Auto-ID for Grammars
Grammar.id was rarely used (it's called once if VERBOSE_GRAMMAR is set), and also there are few grammars where it was non-trivial.  (It was usually just `tag(child)`.)

This made it a good candidate for the headerID treatment, where instead of having to implement a custom `.id()` method inside every Grammar, we use the dynamic nature of JavaScript to introspect into the object and construct a reasonable ID automatically.