Unifying grammarID and cellID
Parsed cells used to be their own hierarchy, but now they're just Grammars.  This means that cellID and grammarID both basically do the same thing now -- turn grammars into strings -- just with a different output syntax.

grammarID is much more flexible, but for the limited world of cells it's a bit verbose in places, so this commit splits the difference and puts some of the cellID abbreviations into grammarID (like a literal on the dummy tape isn't `$T:hello`, it's just `hello`).