getChildren() now introspective
Rather than have different implementations of Grammar.getChildren() for every arity and special case of Grammar, we can just have one.  Every implementation could be generalized as "every instance of Grammar or Grammar[] that's in Object.values(this)".  We can do that introspectively rather than explicitly, by looping through the values and testing what they are, just like mapChildren or grammarID.

This is actually more LoC than before, but it's easier to reason about because it's not tied up in the specific inheritance hierarchy of Grammars.  (E.g., you don't have to remember whether, say, EmbedGrammar is a UnaryGrammar to reason about how it acts here.)  Actually, getting the variably numbered and named children was the main remaining thing the Grammar hierarchy actually handled; maybe we can get rid of AtomicGrammar, UnaryGrammar, etc.