grammarToID -> toStr
(1) `headerID` was doing basically the same thing as `grammarToID`, just in a less general way (because of the more limited complexity of `Header`s.  `grammarToID` was already written so that it'd basically work for any component, but had a few typechecking snags if you actually used it that way, so this fixes that.

(2) It wasn't quite right to say that these are "IDs", they're not necessarily unique and can't serve to identify a particular grammar.  They're just strings and we really only use them for logging.  So now this is just called `toStr()`.