First attempt at static Grammar.tapes
The tape structure of Grammars had always been ephemeral -- it changed as the grammar changed, was lost after cloning, and needed to be recalculated frequently.  But to reason about compilation, we need to be able to reason more solidly about what tapes exist and when.

This is a first attempt at making tapes a permanent member of Grammars, calculated early and kept throughout the passes.

It works in a different way -- more like type inference in a general-purpose language.

(1) We collect the tape information for each identifier (e.g. that identifier X has `{text,gloss}`).  Instead of recursing when we reach an Embed, we just include it as a special variable tape identifier (`tapeRef`) that will be resolved later.

(2) We take all the resulting <identifier,tapes> pairs and perform unification, e.g. if X has `{a,b}` and Y has `{c,${X}}`, we resolve that to Y = `{a,b,c}`.  By the end, only literal tapes remain.

(3) We feed this information back through the Grammar tree so that tapeRefs can be replaced by literal tapes.

This avoids a lot of tree re-traversal since we don't follow Embeds, and lets us potentially catch certain tape paradoxes that the old method was blind to.  (E.g., a recursive Embed inside a Rename that renames one of its tapes. We've put off actually dealing with those because we didn't have a method to detect that was happening.)

But also, it complicates the nature of the tapes object we have to maintain during this process.  It can't just be a list of strings some of which are interpreted specially; for step (2) to work it has to also capture potentially-nested renaming structures that may surround Embeds.  (E.g. if X has `{a,b}` and Y has `{c,Rename(${X},a,d)}`, that Rename node is necessary to correctly resolve this to `{c,d,b}` rather than `{c,a,b}`.)

This commit represents a prototype of the system that works (in the sense that it gets the right tapes for a variety of test grammars), but isn't yet fully hooked up to the Passes.  The tricky bit is that any new structure created during a pass/staging needs this step, and that's not quite how we've envisioned passes in the past.